Move more standalone classes and functions into utils.
- Mutex, BiMap and runtime/util.ts (renamed to misc.ts)
- Deleted the unused (and unimplemented) bytes.ts
