+++ src/runtime/storage/direct-store.ts
@@ -13,7 +13,7 @@ import {CRDTModel, CRDTTypeRecord, CRDTChange, ChangeType, CRDTError, CRDTData} 
 import {DriverFactory} from './drivers/driver-factory.js';
 import {Driver} from './drivers/driver.js';
 import {ActiveStore, ProxyCallback, ProxyMessageType, ProxyMessage, StoreConstructorOptions} from './store-interface.js';
-import {noAwait} from '../util.js';
+import {noAwait} from '../../utils/lib-utils.js';
 
 export enum DirectStoreState {Idle = 'Idle', AwaitingResponse = 'AwaitingResponse', AwaitingResponseDirty = 'AwaitingResponseDirty', AwaitingDriverModel = 'AwaitingDriverModel'}
