+++ src/runtime/storage/reference-mode-store.ts
@@ -13,10 +13,9 @@ import {DirectStoreMuxer} from './direct-store-muxer.js';
 import {DirectStore} from './direct-store.js';
 import {StorageKey} from './storage-key.js';
 import {Type, CollectionType, ReferenceType, SingletonType, MuxType, EntityType} from '../../types/lib-types.js';
-import {Producer, Consumer, Runnable, Dictionary} from '../../utils/lib-utils.js';
+import {Producer, Consumer, Runnable, Dictionary, noAwait} from '../../utils/lib-utils.js';
 import {PropagatedException} from '../arc-exceptions.js';
 import {Store} from './store.js';
-import {noAwait} from '../util.js';
 import {SerializedEntity} from '../entity.js';
 import {ReferenceModeStorageKey} from './reference-mode-storage-key.js';
 import {CRDTTypeRecordToType} from './storage.js';
