+++ src/runtime/storage/store-interface.ts
@@ -14,10 +14,9 @@ import {CRDTTypeRecord, CRDTModel} from '../../crdt/lib-crdt.js';
 import {Exists} from './drivers/driver.js';
 import {StorageKey} from './storage-key.js';
 import {StorageProxy} from './storage-proxy.js';
-import {Producer, Dictionary} from '../../utils/lib-utils.js';
+import {Producer, Dictionary, noAwait} from '../../utils/lib-utils.js';
 import {ChannelConstructor} from '../channel-constructor.js';
 import {StorageProxyMuxer} from './storage-proxy-muxer.js';
-import {noAwait} from '../util.js';
 import {AbstractStore} from './abstract-store.js';
 import {CRDTTypeRecordToType, CRDTMuxEntity} from './storage.js';
 import {StoreRecord} from './direct-store-muxer.js';
