+++ src/runtime/tests/wasm-test.ts
@@ -14,7 +14,7 @@ import {EntityType, ReferenceType} from '../../types/lib-types.js';
 import {Entity} from '../entity.js';
 import {Reference} from '../reference.js';
 import {StringEncoder, StringDecoder, DynamicBuffer} from '../wasm.js';
-import {BiMap} from '../bimap.js';
+import {BiMap} from '../../utils/lib-utils.js';
 import {assertThrowsAsync} from '../../testing/test-util.js';
 
 async function setup() {
