+++ src/runtime/wasm.ts
@@ -15,11 +15,10 @@ import {Reference} from './reference.js';
 import {Type, EntityType, CollectionType, ReferenceType, SingletonType, Schema} from '../types/lib-types.js';
 import {Storable} from './storable.js';
 import {Particle} from './particle.js';
-import {Dictionary} from '../utils/lib-utils.js';
+import {Dictionary, BiMap} from '../utils/lib-utils.js';
 import {PECInnerPort} from './api-channel.js';
 import {UserException} from './arc-exceptions.js';
 import {ParticleExecutionContext} from './particle-execution-context.js';
-import {BiMap} from './bimap.js';
 import {CollectionHandle, SingletonHandle, Handle} from './storage/handle.js';
 import {CRDTTypeRecord} from '../crdt/lib-crdt.js';
