+++ src/utils/internal/misc.ts
@@ -8,8 +8,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {assert} from '../platform/assert-web.js';
-import {Dictionary} from '../utils/lib-utils.js';
+import {assert} from '../../platform/assert-web.js';
+import {Dictionary} from '../lib-utils.js';
 
 /**
  * Returns the set delta between two lists based on direct object comparison.
